+++ src/cli/mod.rs
@@ -1,7 +1,5 @@
 pub(crate) mod build;
-pub(crate) mod doctor;
 pub(crate) mod execution;
-pub(crate) mod new_project;
 pub(crate) mod package_lock;
 pub(crate) mod package_resolver;
 pub(crate) mod project;